lead_triage_env/server/requirements.txt
# Minimal server image dependencies (pip-based Docker build)
openenv-core[core]>=0.2.2
fastapi>=0.115.0
pydantic>=2.0.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
